+++ src/arch_armcgt.h
@@ -13,7 +13,6 @@
 #include <time.h>
 
 #define MG_PATH_MAX 100
-#define MG_ENABLE_SOCKET 0
 #define MG_ENABLE_DIRLIST 0
 
 #endif
