+++ tutorials/nxp/frdm-rw612-xpresso-freertos-builtin-wifi/source/mongoose_config.h
@@ -1,7 +1,8 @@
 #pragma once
 
 // See https://mongoose.ws/documentation/#build-options
-#define MG_ARCH MG_ARCH_FREERTOS
+#define MG_ARCH MG_ARCH_ARMGCC
+#define MG_ENABLE_FREERTOS 1
 
 #define MG_ENABLE_TCPIP 1
 #define MG_ENABLE_DRIVER_NXP_WIFI 1
