+++ tutorials/nxp/frdm-rw612-xpresso-freertos-lwip/source/mongoose_config.h
@@ -1,6 +1,7 @@
 #pragma once
 
 // See https://mongoose.ws/documentation/#build-options
-#define MG_ARCH MG_ARCH_FREERTOS
+#define MG_ARCH MG_ARCH_ARMGCC
+#define MG_ENABLE_FREERTOS 1
 
 #define MG_ENABLE_LWIP 1
