--- tutorials/stm32/nucleo-f429zi-cube-freertos-lwip/.settings/com.st.stm32cube.ide.mcu.sfrview.prefs
@@ -1,2 +1,0 @@
-eclipse.preferences.version=1
-sfrviewstate={"fFavorites"\:{"fLists"\:{}},"fProperties"\:{"fNodeProperties"\:{}}}
