--- tutorials/stm32/nucleo-f429zi-cube-freertos-lwip/.settings/org.eclipse.cdt.core.prefs
@@ -1,6 +1,0 @@
-doxygen/doxygen_new_line_after_brief=true
-doxygen/doxygen_use_brief_tag=false
-doxygen/doxygen_use_javadoc_tags=true
-doxygen/doxygen_use_pre_tag=false
-doxygen/doxygen_use_structural_commands=false
-eclipse.preferences.version=1
