--- tutorials/stm32/nucleo-f429zi-cube-freertos-lwip/.settings/stm32cubeide.project.prefs
@@ -1,5 +1,0 @@
-635E684B79701B039C64EA45C3F84D30=1767C1FD5A90CBA93D279BE91F305A67
-66BE74F758C12D739921AEA421D593D3=0
-8DF89ED150041C4CBC7CB9A9CAA90856=7FFD126C0666B803FF7CCACF304BE6C0
-DC22A860405A8BF2F2C095E5B6529F12=7FFD126C0666B803FF7CCACF304BE6C0
-eclipse.preferences.version=1
