+++ tutorials/stm32/nucleo-f429zi-cube-freertos-lwip/Core/Inc/mongoose_config.h
@@ -9,7 +9,8 @@
 
 #define MG_ENABLE_CUSTOM_MILLIS  0
 #define MG_ENABLE_MBEDTLS  0
-#define MG_ARCH MG_ARCH_FREERTOS
+#define MG_ARCH MG_ARCH_ARMGCC
+#define MG_ENABLE_FREERTOS 1
 #define MG_STMPACK_NET 1
 #define MG_ENABLE_CUSTOM_RANDOM  1
