--- tutorials/stm32/nucleo-f746zg-cube-freertos-lwip/.settings/stm32cubeide.project.prefs
@@ -1,5 +1,0 @@
-635E684B79701B039C64EA45C3F84D30=7EC62A8EEAB33E6B6251E64B7094FD8B
-66BE74F758C12D739921AEA421D593D3=1
-8DF89ED150041C4CBC7CB9A9CAA90856=B51B08583BFB231F447A5555C047C200
-DC22A860405A8BF2F2C095E5B6529F12=B51B08583BFB231F447A5555C047C200
-eclipse.preferences.version=1
