+++ tutorials/stm32/nucleo-f746zg-make-freertos-tcp/mongoose_config.h
@@ -1,9 +1,9 @@
 #pragma once
 
 // See https://mongoose.ws/documentation/#build-options
-#define MG_ARCH MG_ARCH_FREERTOS
+#define MG_ARCH MG_ARCH_ARMGCC
+#define MG_ENABLE_FREERTOS 1
 
 #define MG_ENABLE_FREERTOS_TCP 1
-#define MG_ENABLE_DRIVER_STM32F 1
 #define MG_ENABLE_CUSTOM_RANDOM 1
